+++ bot/resource/cog/registration/factory/force_register_student_request_factory.py
@@ -6,8 +6,8 @@ from bot.application.student.exceptions.invalid_discord_id_format_exception impo
 from bot.application.student.exceptions.invalid_ni_format_exception import (
     InvalidNIFormatException,
 )
-from bot.application.student.validators.discord_id_validator import DiscordIdValidator
-from bot.application.student.validators.ni_validator import NIValidator
+from bot.resource.cog.validators.discord_id_validator import DiscordIdValidator
+from bot.resource.cog.validators.ni_validator import NIValidator
 from bot.domain.student.attribut.discord_user_id import DiscordUserId
 from bot.domain.student.factory.ni_factory import NIFactory
 from bot.resource.cog.registration.factory.request_factory import RequestFactory
