+++ bot/resource/cog/registration/factory/unregister_student_request_factory.py
@@ -3,7 +3,7 @@ from typing import Tuple
 from bot.application.student.exceptions.invalid_discord_id_format_exception import (
     InvalidDiscordIdFormatException,
 )
-from bot.application.student.validators.discord_id_validator import DiscordIdValidator
+from bot.resource.cog.validators.discord_id_validator import DiscordIdValidator
 from bot.domain.student.attribut.discord_user_id import DiscordUserId
 from bot.resource.cog.registration.factory.request_factory import RequestFactory
 from bot.resource.cog.registration.request.unregister_student_request import (
