+++ bot/resource/cog/validators/name_validator.py
